moved some widgets to widgets.py
